'Add' methods to db stub repos. GetProjectSiblingFolder method

+++ BackEnd/WorkflowApp/InitializeFileTestData.cs
@@ -1,12 +1,29 @@
 using System.IO;
 using GM.ProcessRecording;
+using GM.FileDataRepositories;
 
 namespace GM.Workflow
 {
     public static class InitializeFileTestData {
-        public static void CopyTestData(string testfilesPath, string datafilesPath)
+        public static string CopyTestData(string testfilesPath, string datafilesPath, bool deleteProcessing)
         {
-            Directory.CreateDirectory(datafilesPath + "\\" + "RECEIVED");
+            if (!Directory.Exists(datafilesPath))
+            {
+                Directory.CreateDirectory(datafilesPath);
+                Directory.CreateDirectory(datafilesPath + "/RECEIVED");
+                Directory.CreateDirectory(datafilesPath + "/PROCESSING");
+                Directory.CreateDirectory(datafilesPath + "/COMPLETED");
+            } else {
+                if (deleteProcessing)
+                {
+                    GMFileAccess.DeleteDirectoryContents(datafilesPath + @"\PROCESSING");
+                }
+            }
+
+            if (!Directory.Exists(testfilesPath)){
+                Directory.CreateDirectory(testfilesPath);
+                return "TESTDATA folder missing";
+            }
 
             // These are the test files that we will copy.
             string[] files = new string[]
@@ -15,7 +32,6 @@ namespace GM.Workflow
                //  as meeting #4 and status = "Received"
                "USA_PA_Philadelphia_Philadelphia_CityCouncil_en_2017-12-07.pdf",
 
-               // TODO - This is about 58MB. Github prefers < 50.
                // This meeting exists in MeetingRepository_Stub.cs
                //  as meeting #5 and status = "Received"
                "USA_ME_LincolnCounty_BoothbayHarbor_Selectmen_en_2017-01-09.mp4",
@@ -26,24 +42,32 @@ namespace GM.Workflow
                 "USA_ME_LincolnCounty_BoothbayHarbor_Selectmen_en_2017-02-15.mp4"
             };
 
+            if (files.Length == 0)
+            {
+                return "TESTDATA folder empty";
+            }
+
             foreach (string file in files)
             {
-                string destination = datafilesPath + "\\" + "RECEIVED" + "\\" + file;
-                if (!File.Exists(destination))
-                {
-                    string source = testfilesPath + "\\" + file;
+                string source = testfilesPath + "\\" + file;
 
-                    // For testing, use only the first 9 minutes of the video recordings.
-                    if (file.EndsWith(".mp4"))
+                if (File.Exists(source)) {
+                    string destination = datafilesPath + "\\" + "RECEIVED" + "\\" + file;
+                    if (!File.Exists(destination))
                     {
-                        SplitRecording splitRecording = new SplitRecording();
-                        splitRecording.ExtractPart(source, destination, 0, 540);  // 9 * 60 sec.
-                    } else
-                    {
-                        File.Copy(source, destination);
+                        // For testing, use only the first 9 minutes of the video recordings.
+                        if (file.EndsWith(".mp4"))
+                        {
+                            SplitRecording splitRecording = new SplitRecording();
+                            splitRecording.ExtractPart(source, destination, 0, 540);  // 9 * 60 sec.
+                        } else
+                        {
+                            File.Copy(source, destination);
+                        }
                     }
                 }
             }
+        return null;
         }
     }
 }
